requirements.txt
amqp==5.1.1
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
billiard==4.1.0
bleach==6.0.0
blinker==1.6.2
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==14.3
comm==0.1.2
contourpy==1.0.7
cryptography==41.0.2
cycler==0.11.0
daq==0.4.2
dash==2.10.2
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
dash-testing-stub==0.0.2
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
diskcache==5.6.1
distlib==0.3.7
docopt==0.6.2
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.2
Flask==2.2.5
fonttools==4.39.0
fqdn==1.5.1
fs==2.4.16
h11==0.14.0
humanfriendly==10.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
kombu==5.3.1
lxml==4.9.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
multiprocess==0.70.14
mysql-connector-python==8.1.0
mysql-connector-python-rf==2.2.2
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
percy==2.0.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.10.0
plotly==5.15.0
pluggy==1.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyrsistent==0.19.3
pyserial==3.5
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
redis==4.6.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
selenium==4.2.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
stack-data==0.6.2
tenacity==6.3.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.16
urllib3-secure-extra==0.1.0
verboselogs==1.7
vine==5.0.0
virtualenv==20.24.2
waitress==2.1.2
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
wsproto==1.2.0
